Fixed some typos in the documentation
Moved permission_error and policy_error into the correct files.
